pyproject.toml
[project]
name = "truepay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.13",
    "cryptography>=45.0.6",
    "django==5.2.5",
    "django-allauth>=65.11.1",
    "django-debug-toolbar>=6.0.0",
    "django-storages>=1.14.6",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.0",
    "pycryptodome>=3.23.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.0",
    "pyotp>=2.9.0",
    "qrcode[pil]>=7.4.2",
    "requests>=2.32.5",
    "resend>=2.13.0",
    "openpyxl>=3.1.5",
    "django-celery-beat>=2.8.1",
    "celery>=5.3.0",
    "djangorestframework>=3.16.1",
    "whitenoise>=6.10.0",
    "rabbitmq>=0.2.0",
]
